+++ qtwebengine/src/3rdparty/chromium/components/safe_browsing/content/browser/web_api_handshake_checker.cc
@@ -7,6 +7,7 @@
 #include "components/safe_browsing/content/browser/web_ui/safe_browsing_ui.h"
 #include "components/safe_browsing/core/browser/safe_browsing_url_checker_impl.h"
 #include "components/safe_browsing/core/browser/url_checker_delegate.h"
+#include "components/safe_browsing/core/common/features.h"
 #include "content/public/browser/browser_task_traits.h"
 #include "content/public/browser/browser_thread.h"
 #include "content/public/browser/web_contents.h"
@@ -14,10 +15,10 @@
 
 namespace safe_browsing {
 
-class WebApiHandshakeChecker::CheckerOnIO
-    : public base::SupportsWeakPtr<WebApiHandshakeChecker::CheckerOnIO> {
+class WebApiHandshakeChecker::CheckerOnSB
+    : public base::SupportsWeakPtr<WebApiHandshakeChecker::CheckerOnSB> {
  public:
-  CheckerOnIO(base::WeakPtr<WebApiHandshakeChecker> handshake_checker,
+  CheckerOnSB(base::WeakPtr<WebApiHandshakeChecker> handshake_checker,
               GetDelegateCallback delegate_getter,
               const GetWebContentsCallback& web_contents_getter,
               int frame_tree_node_id)
@@ -37,7 +38,9 @@ class WebApiHandshakeChecker::CheckerOnIO
   }
 
   void Check(const GURL& url) {
-    DCHECK_CURRENTLY_ON(content::BrowserThread::IO);
+    DCHECK_CURRENTLY_ON(base::FeatureList::IsEnabled(kSafeBrowsingOnUIThread)
+                            ? content::BrowserThread::UI
+                            : content::BrowserThread::IO);
     DCHECK(delegate_getter_);
     DCHECK(web_contents_getter_);
 
@@ -51,10 +54,7 @@ class WebApiHandshakeChecker::CheckerOnIO
             /*render_frame_id=*/MSG_ROUTING_NONE,
             /*originated_from_service_worker=*/false);
     if (skip_checks) {
-      OnCompleteCheck(/*slow_check=*/false, /*proceed=*/true,
-                      /*showed_interstitial=*/false,
-                      /*did_perform_real_time_check=*/false,
-                      /*did_check_allowlist=*/false);
+      OnCompleteCheckInternal(/*proceed=*/true);
       return;
     }
 
@@ -64,18 +64,20 @@ class WebApiHandshakeChecker::CheckerOnIO
         url_checker_delegate, web_contents_getter_,
         /*render_process_id=*/content::ChildProcessHost::kInvalidUniqueID,
         /*render_frame_id=*/MSG_ROUTING_NONE, frame_tree_node_id_,
-        /*real_time_lookup_enabled=*/false,
-        /*can_rt_check_subresource_url=*/false,
+        /*url_real_time_lookup_enabled=*/false,
+        /*can_urt_check_subresource_url=*/false,
         /*can_check_db=*/true, /*can_check_high_confidence_allowlist=*/true,
         /*url_lookup_service_metric_suffix=*/".None", last_committed_url_,
         content::GetUIThreadTaskRunner({}),
         /*url_lookup_service=*/nullptr, WebUIInfoSingleton::GetInstance(),
         /*hash_realtime_service_on_ui=*/nullptr,
         /*mechanism_experimenter=*/nullptr,
-        /*is_mechanism_experiment_allowed=*/false);
+        /*is_mechanism_experiment_allowed=*/false,
+        /*hash_realtime_selection=*/
+        hash_realtime_utils::HashRealTimeSelection::kNone);
     url_checker_->CheckUrl(
         url, "GET",
-        base::BindOnce(&WebApiHandshakeChecker::CheckerOnIO::OnCheckUrlResult,
+        base::BindOnce(&WebApiHandshakeChecker::CheckerOnSB::OnCheckUrlResult,
                        base::Unretained(this)));
   }
 
@@ -85,32 +87,41 @@ class WebApiHandshakeChecker::CheckerOnIO
       SafeBrowsingUrlCheckerImpl::NativeUrlCheckNotifier* slow_check_notifier,
       bool proceed,
       bool showed_interstitial,
-      bool did_perform_real_time_check,
-      bool did_check_allowlist) {
-    DCHECK_CURRENTLY_ON(content::BrowserThread::IO);
+      SafeBrowsingUrlCheckerImpl::PerformedCheck performed_check,
+      bool did_check_url_real_time_allowlist) {
+    DCHECK_CURRENTLY_ON(base::FeatureList::IsEnabled(kSafeBrowsingOnUIThread)
+                            ? content::BrowserThread::UI
+                            : content::BrowserThread::IO);
     if (!slow_check_notifier) {
-      OnCompleteCheck(/*slow_check=*/false, proceed, showed_interstitial,
-                      did_perform_real_time_check, did_check_allowlist);
+      OnCompleteCheckInternal(proceed);
       return;
     }
 
     *slow_check_notifier =
-        base::BindOnce(&WebApiHandshakeChecker::CheckerOnIO::OnCompleteCheck,
+        base::BindOnce(&WebApiHandshakeChecker::CheckerOnSB::OnCompleteCheck,
                        base::Unretained(this), /*slow_check=*/true);
   }
 
-  void OnCompleteCheck(bool slow_check,
-                       bool proceed,
-                       bool showed_interstitial,
-                       bool did_perform_real_time_check,
-                       bool did_check_allowlist) {
-    DCHECK_CURRENTLY_ON(content::BrowserThread::IO);
-    content::GetUIThreadTaskRunner({})->PostTask(
-        FROM_HERE,
-        base::BindOnce(&WebApiHandshakeChecker::OnCompleteCheck,
-                       handshake_checker_, slow_check, proceed,
-                       showed_interstitial, did_perform_real_time_check,
-                       did_check_allowlist));
+  void OnCompleteCheck(
+      bool slow_check,
+      bool proceed,
+      bool showed_interstitial,
+      SafeBrowsingUrlCheckerImpl::PerformedCheck performed_check,
+      bool did_check_url_real_time_allowlist) {
+    OnCompleteCheckInternal(proceed);
+  }
+
+  void OnCompleteCheckInternal(bool proceed) {
+    DCHECK_CURRENTLY_ON(base::FeatureList::IsEnabled(kSafeBrowsingOnUIThread)
+                            ? content::BrowserThread::UI
+                            : content::BrowserThread::IO);
+    if (base::FeatureList::IsEnabled(kSafeBrowsingOnUIThread)) {
+      handshake_checker_->OnCompleteCheck(proceed);
+    } else {
+      content::GetUIThreadTaskRunner({})->PostTask(
+          FROM_HERE, base::BindOnce(&WebApiHandshakeChecker::OnCompleteCheck,
+                                    handshake_checker_, proceed));
+    }
   }
 
   base::WeakPtr<WebApiHandshakeChecker> handshake_checker_;
@@ -126,31 +137,33 @@ WebApiHandshakeChecker::WebApiHandshakeChecker(
     const GetWebContentsCallback& web_contents_getter,
     int frame_tree_node_id) {
   DCHECK_CURRENTLY_ON(content::BrowserThread::UI);
-  io_checker_ = std::make_unique<CheckerOnIO>(
+  sb_checker_ = std::make_unique<CheckerOnSB>(
       weak_factory_.GetWeakPtr(), std::move(delegate_getter),
       web_contents_getter, frame_tree_node_id);
 }
 
 WebApiHandshakeChecker::~WebApiHandshakeChecker() {
   DCHECK_CURRENTLY_ON(content::BrowserThread::UI);
-  content::GetIOThreadTaskRunner({})->DeleteSoon(FROM_HERE,
-                                                 std::move(io_checker_));
+  if (!base::FeatureList::IsEnabled(kSafeBrowsingOnUIThread)) {
+    content::GetIOThreadTaskRunner({})->DeleteSoon(FROM_HERE,
+                                                   std::move(sb_checker_));
+  }
 }
 
 void WebApiHandshakeChecker::Check(const GURL& url, CheckCallback callback) {
   DCHECK_CURRENTLY_ON(content::BrowserThread::UI);
   DCHECK(!check_callback_);
   check_callback_ = std::move(callback);
-  content::GetIOThreadTaskRunner({})->PostTask(
-      FROM_HERE, base::BindOnce(&WebApiHandshakeChecker::CheckerOnIO::Check,
-                                io_checker_->AsWeakPtr(), url));
+  if (base::FeatureList::IsEnabled(kSafeBrowsingOnUIThread)) {
+    sb_checker_->Check(url);
+  } else {
+    content::GetIOThreadTaskRunner({})->PostTask(
+        FROM_HERE, base::BindOnce(&WebApiHandshakeChecker::CheckerOnSB::Check,
+                                  sb_checker_->AsWeakPtr(), url));
+  }
 }
 
-void WebApiHandshakeChecker::OnCompleteCheck(bool slow_check,
-                                             bool proceed,
-                                             bool showed_interstitial,
-                                             bool did_perform_real_time_check,
-                                             bool did_check_allowlist) {
+void WebApiHandshakeChecker::OnCompleteCheck(bool proceed) {
   DCHECK_CURRENTLY_ON(content::BrowserThread::UI);
   DCHECK(check_callback_);
